(adaptors) add posthog visio adaptor
* (adaptors) add posthog base adaptor

Several of our stats are coming from PostHog, so we need to add
a base adaptor to handle them. We here add support for a basic posthog
trend indicator that can be used to fetch the latest value
from a particular posthog insight.

* (adaptors) add visio adaptor

Get monthly active users from posthog instead of getting them from
ProConnect, as most visio users are not logged in during a visio.